test(server): constrain mocked signal contexts
